sandbox: rename class-anchored files to PascalCase
Rename state.ts -> StateBackend.ts and protocol.ts -> BackendProtocol.ts
to match the packages/components convention (PascalCase for files whose
primary export is a same-named class/interface).
